moves polystore models into models module
