+++ tdm_ingestion/ingestion.py
@@ -2,43 +2,9 @@
 
 import uuid
 from abc import ABC, abstractmethod
-from datetime import datetime, timezone
 from typing import List
 
-
-class Measure(ABC):
-    @abstractmethod
-    def to_dict(self) -> dict:
-        pass
-
-
-class ValueMeasure(Measure):
-    def __init__(self, value):
-        self.value = value
-
-    def to_dict(self) -> dict:
-        return {'value': self.value}
-
-
-class RefMeasure(Measure):
-    def __init__(self, ref, index):
-        self.ref = ref
-        self.index = index
-
-    def to_dict(self) -> dict:
-        return {'reference': self.ref, 'index': self.index}
-
-
-class TimeSeries:
-    def __init__(self, utc_time: datetime, sensorcode: uuid.UUID, measure: Measure):
-
-        self.time = utc_time
-        self.sensorcode = sensorcode
-        self.measure = measure
-
-    def to_dict(self, time_format: str = '%Y-%m-%dT%H:%M:%SZ') -> dict:
-        return {'time': self.time.strftime(time_format), 'sensorcode': str(self.sensorcode),
-                'measure': self.measure.to_dict()}
+from tdm_ingestion.models import TimeSeries
 
 
 class Message:
@@ -50,7 +16,8 @@ class Message:
 class Consumer(ABC):
 
     @abstractmethod
-    def poll(self, timeout_s: int = -1, max_records: int = -1) -> List[Message]:
+    def poll(self, timeout_s: int = -1, max_records: int = -1) -> List[
+        Message]:
         pass
 
 
@@ -70,13 +37,15 @@ class MessageConverter(ABC):
 
 
 class Ingester:
-    def __init__(self, consumer: Consumer, storage: Storage, converter: MessageConverter):
+    def __init__(self, consumer: Consumer, storage: Storage,
+                 converter: MessageConverter):
         self.consumer = consumer
         self.storage = storage
         self.converter = converter
 
     def process(self, timeout_s: int = -1, max_records: int = 1):
-        self.storage.write(self.converter.convert(self.consumer.poll(timeout_s, max_records)))
+        self.storage.write(
+            self.converter.convert(self.consumer.poll(timeout_s, max_records)))
 
 
 if __name__ == '__main__':
@@ -111,8 +80,10 @@ if __name__ == '__main__':
     with open(args.conf_file, 'r') as conf_file:
         conf = yaml.safe_load(conf_file)
         logging.debug('conf %s', conf)
-        storage = import_class(conf['storage']['class'])(**conf['storage']['args'])
-        consumer = import_class(conf['consumer']['class'])(**conf['consumer']['args'])
+        storage = import_class(conf['storage']['class'])(
+            **conf['storage']['args'])
+        consumer = import_class(conf['consumer']['class'])(
+            **conf['consumer']['args'])
         ingester_process_args = conf['ingester']['process']
 
     ingester = Ingester(consumer, storage, NgsiConverter())
